==> 04_report.Rmd <==
---
title: "Data analysis support for TNC-NTRI rangeland management project publication"
author: "Stefan Gehrig, <a href='https://www.estimact.com'>estimact</a>"
date: "`r format(Sys.time(), '%d %B %Y')`"
subtitle: "Version 6"
output: 
  bookdown::html_document2:
    toc: TRUE
    toc_float: TRUE
    toc_collapsed: TRUE
    toc_depth: 4
    number_sections: FALSE
    css: utils/style.css
    includes:
      after_body: utils/footer.html
bibliography: utils/refs.bib
---

<style>
    .table td, .table th {
        font-size: 10px;
    }
</style>

<!-- START OF -->
<br>

***
<!-- DOCUMENT -->

```{r setup, include=FALSE}
library(knitr)
library(kableExtra)
library(tidyverse)
library(brms)
library(dagitty)
library(ggdag)
knitr::opts_chunk$set(echo = FALSE)

summary_tier3 <- readRDS("outputs/summary_tier3.rds")
irm_items = readRDS("outputs/irm_items.rds")
irm_items_dist = readRDS("outputs/irm_items_dist.rds")
fit = readRDS("outputs/fit_ord_cum_2pl.rds")
fit2 = readRDS("outputs/fit_bare.rds")
```

## Villages in the sample

Figure \@ref(fig:map) shows a map of the villages included in the analysis.

```{r map, fig.align='center', fig.cap='Map of villages in the sample and Tanzanian administrative boundaries.', out.width="80%"}
include_graphics("outputs/vilmap.png")
```
<br>

## Data processing and description

The raw data can be browsed under <https://stefgehrig.shinyapps.io/TNC-NTRI/>. Starting with those data, all Diagnostic Framework (DF) variables were transformed to be on the same scale from 1 to 3 and to only take integer values for the presented initial analyses. So the procedure was:

 - All variables which had already originally been coded as ordinal with response options 1, 2, 3 were not changed (irrespective of whether the whole range of responses actually occurred in the data)
 - Binary variables (response options 1, 2), were set from 1, 2 to 1, 3 in order to have the same numerical midpoint of their range
 - Continuous variables (e.g., Number of actors, Change in population size, Economic heterogeneity, Mean annual rainfall, ...) were linearly transformed such that the minimum was 1 and the maximum was 3 and then rounded to the nearest integer (1, 2, or 3). **This was only done for the initial exploration in this subsection, but their more fine-grained scale will be respected during modeling.**

The meaning of code directions is shown in Table \@ref(tab:sds). The coding was such that -- where such a statement could sensibly be made -- higher values refer to better governance/outcomes/conditions for sustainability etc.

<p>Note that some of the Diagnostic Framework (DF) variables show no variation between villages in the sample. See Table \@ref(tab:sds) with means and standard deviations (SD) for each DF variable. Variables which show no variation are highlighted in <span style="color:#FF7F50">orange</span>.</p>

```{r sds}
color.me <- which(summary_tier3$`SD (rescaled)` == 0)
summary_tier3 %>% 
  mutate(across(.cols = c(`Mean (rescaled)`, `SD (rescaled)`), ~round(.x,2))) %>% 
  kable(
    format = "html",
    caption = "Mean and SD for Tier 3 variables and their coding.") %>% 
  kable_styling() %>%
  row_spec(color.me, background = "coral")
```

In Figure \@ref(fig:radar), one radar plot per category of the diagnostic framework (Tier 1) and per village is shown. This descriptive overview allows to get a better sense of the variation we have and which DF variables and categories might be more fruitful to analyze in more depth. The reader is referred to Table \@ref(tab:sds) for interpreting the scores (i.e., what does a low value of 1 [minimum] and a high value of 3 [maximum] mean?).

```{r radar, fig.align='center', fig.cap='Radar plots of scores for all DF variables. Columns are Tier-1 framework categories, rows are villages.', out.width="100%"}
include_graphics("outputs/spiders_separate.png")
```

## Initial data exploration

For an initial exploration, I only looked at specific bivariate associations between variables, driven by some of previously discussed research questions. Average scores for  `Governance Systems (GS)` , `Interactions (I)` and `Outcomes (O)` DF variables were created for each village. Higher scores on `Governance Systems (GS)` and `Interactions (I)` can be interpreted as more developed governance systems and higher scores on `Outcomes (O)` can be interpreted as better commons outcomes (see Table \@ref(tab:sds)).

For two of the villages, it was found in the qualitative data sources that responses about common lands during interviews and focus groups might not have been reliable (see e-mail 2023-05-22 from Monique). These are marked with red circles in all plots that show individual data points (in case qualitative data sources are involved).

The association is shown in Figure \@ref(fig:scattergso).

```{r scattergso, fig.align='center', fig.cap='Scatter plot of village-level averages of scores in Governance Systems (GS), Interactions (I) and Outcomes (O) variables. Pearson correlation coefficient is shown.', out.width="80%"}
include_graphics("outputs/scatter_gs_i_o.png")
```
<br>

A positive association is clearly indicated. We can also look at the correlations of `Governance Systems (GS)` and `Outcomes (O)` variables on a variable-by-variable basis, as in the correlation matrix in Figure \@ref(fig:corrgso).

```{r corrgso, fig.align='center', fig.cap='Pearson correlation coefficients between single Governance Systems (GS) and Outcomes (O) variables.', out.width="70%"}
include_graphics("outputs/corrm_gs_o.png")
```
<br>

The outcome variable `O2.3: Invasives` seems to be least strongly associated with the `Governance Systems (GS)` variables.

A natural follow-up question would be to see how these relationships look after adjusting for rainfall. We would probably expect that we gain certainty about the positive relationship between resource governance and commons outcomes when partialling out the effect of natural variation in climate between villages. Therefore, I residualized the three `Outcomes (O)` variables via a linear model with `ECO1.1: Rainfall patterns` (average of mean annual precipitation 2016-2020 in mm) as a predictor before correlating them with the `Governance Systems (GS)` variables again. Technically, this estimates a semi-partial correlation. Interestingly, the relationships get weaker, as shown in Figure \@ref(fig:corrgso2).

```{r corrgso2, fig.align='center', fig.cap='Pearson correlation coefficients between single Governance Systems (GS) and Outcomes (O) variables, with the effect of rainfall on Outcomes (O) variables partialled out.', out.width="70%"}
include_graphics("outputs/corrm_gs_o_residualizedrain.png")
```
<br>

Digging a bit deeper, this seems to be because rainfall is strongly negatively correlated both with the `Governance Systems (GS)` and `Outcomes (O)` variables (see Figure \@ref(fig:allcorrs)). Therefore, it competes with `Governance Systems (GS)` in explaining variation in `Outcomes (O)` variables.

This is confirmed in a Principal Component Analysis on all averaged DF dimensions (Figure \@ref(fig:pca)).

```{r pca, fig.align='center', fig.cap='Biplot from PCA on average scores of villages in each DF dimension.', out.width="95%"}
include_graphics("outputs/pca_descr.png")
```

We can see in Figure \@ref(fig:pca) that ...

- (again) the `Outcomes (O)` responses correlate positively with, for example, the `Governance Systems (GS)` responses (see the color coding in green)
- The Tier-1 DF variables are correlated in the sense that "positive" responses (in the sense that they describe favorable governance conditions and processes) in `Governance Systems (GS)`, for example, are associated with more "positive" responses in `Actors (A)` and `Interactions (I)` (see the respective arrows pointing in a similar direction)
- Two dimensions can explain a lot of variation (more than 80\%, **but note** that this assessment is based on averaged scores per dimension, so a lot of variability is ignored in this analysis)
- (again) Governance and rainfall are substantially negatively correlated: Where there is more rainfall, there are "lower" responses in the `Governance Systems (GS)` variable and related variables, like `Actors (A)` or `Interactions (I)`

The latter is a peculiarity that should be recognized and *might* be due to confounding by factors directly or indirectly related to geography: The more Northern villages, close to the Kenyan border (e.g., Matale A or Ngoswaki, see Figure \@ref(fig:map)) experience less rainfall, as shown in Figure \@ref(fig:vilrain) below. These are also the villages which tend to achieve higher scores in DF variables related to governance compared to villages more in the interior of the country like Kimana (as can be seen in the Biplot in Figure \@ref(fig:pca)), where rainfall is more abundant (Figure \@ref(fig:vilrain)). It is currently unclear how we can best explain this.

```{r vilrain, fig.align='center', fig.cap='Average annual rainfall over time for all study villages.', out.width="80%"}
include_graphics("outputs/vilrain.png")
```

In a next step, the GIS-based measurements of land use practices are introduced. First, Figure \@ref(fig:gisbased) just shows a scatter plot of the two GIS-based outcome trend measures and two reference lines, which indicate zero change. We see that (i) crop land has increased in all villages (all values $>0$) and that there is generally a negative association: Villages that have experienced stronger increases in crop land have also experienced a stronger reduction in bare ground.

```{r gisbased, fig.align='center', fig.cap='Scatter plot of the two GIS-based outcome trend measures, along with reference lines which indidate no change. Pearson correlation coefficient is shown.', out.width="50%"}
include_graphics("outputs/scatter_gis.png")
```

I estimate a factor change in bare ground and crop land, such that, for example, a value of 1 means that the proportion of village area which is bare ground/crop land did not change from the pre- to the post-period, whereas a value of 1.5 means that the proportion of village area which is bare ground or crop land increased by 50\%. This ratio is log-transformed, maintaining the interpretation that larger values signify an increase in bare ground or crop land, respectively. Positive values hence mean that the proportion of land of the respective type has increased.

There are no clear associations between average score in `Governance Systems (GS)` and the GIS-based measure of change in bare land, but a slight positive tendency between average score in `Governance Systems (GS)` and change in crop land. (Figure \@ref(fig:scattergis)).

```{r scattergis, fig.align='center', fig.cap='Scatter plot of village-level averages of scores in Governance Systems (GS) and GIS-based outcome measures. Pearson correlation coefficient is shown.', out.width="90%"}
include_graphics("outputs/scatter_gs_bare_crop.png")
```

Via a simple linear model, I residualize the GIS-based outcome measure on rainfall, therefore adjusting it for differences in average precipitation between villages. The associations become stronger, as can be seen in Figure \@ref(fig:scattergisres). Increases in bare ground were on average more successfully prevented in village with high scores, controlling for rainfall.

```{r scattergisres, fig.align='center', fig.cap='Scatter plot of village-level averages of scores in Governance Systems (GS) and GIS-based outcome measures, adjusted for precipitation. Pearson correlation coefficient is shown.', out.width="90%"}
include_graphics("outputs/scatter_gs_bare_crop_resid.png")
```

How do the GIS-based outcome trend measures relate to the DF variable `O2.1: Commons condition trend`, which is coded from the qualitative data sources? Figure \@ref(fig:boxcommon) shows that there is no obvious association.

```{r boxcommon, fig.align='center', fig.cap='Box plot of the GIS-based outcome trend measures, stratified by village score for the variable O2.1: Commons condition trend.', out.width="70%"}
include_graphics("outputs/boxplots_commen_trend.png")
```

For completeness, I also show all pairwise correlations among all DF variables and GIS-based outcome trend measures in Figure \@ref(fig:allcorrs).

```{r allcorrs, fig.align='center', fig.cap='Pairwise Pearson correlation coefficients among all DF variables and GIS-based outcome trend measures.', out.width="100%"}
include_graphics("outputs/corrm_all_lup.png")
```
<br>

## Item Response Model

The previous analyses were ad hoc to describe some tendencies in the data. We now follow a more stringent approach with better theoretical justification, more accounting for uncertainty, and more targeted research question(s). The workhorse of the analysis will be a Bayesian Item Response Model (IRM) fit with `brms` [@bürkner2020bayesian]. In what follows, we call the Tier-3 DF variables based on qualitative data collection "items".

In brief, the model assumes that there is underlying latent variable which affects the ordinal responses obtained in the interviews and focus group discussions and that the value of this latent variable, in turn, differs across villages and items. Further, items are allowed to differ in their discrimination. This should be accounted for, since there is quite a difference in the between-village variability of responses among items (Table \@ref(tab:sds)). So, some items have high, others have low discriminatory power.

The model is hierarchical in the sense that is uses adaptive priors and partial pooling for both item effects and village effects. Given the relative abundance of parameters and only few independently observed villages, weakly informative Bayesian priors are used which shrink item and village effects towards zero. This ensures model convergence and stability of estimates, thereby also being generally conservative.

Note that the approach allows us to keep the items with no variability (in the sample!) in the analysis. We use a cumulative logit link, assuming common slopes (i.e., proportional odds).  Hence, in the framework of IRM, the model could be called a **2PL graded responses model** [@bürkner2020bayesian].

The central feature that makes this model type attractive is that it allows to obtain a posterior for the correlation of different groups of items. By specifying two item groups *Governance* and *Outcomes*, it is possible to estimate the correlation of the two underlying latent variables for Governance conditions/processes and Commons outcomes **without falling back on ad hoc methods like averaging**, which discards uncertainty and assumes interval scaling of the ordinal variables. Of course, we should only model variables obtained in qualitative data collection as items (and not, for example, rainfall). Therefore, the first model presented below does not include any variables measured *not* via ordinal or binary responses in interviews or focus groups.

As items going into the *Governance* items group, we choose all DF variables in the `Governance Systems (GS)` category, but also others (based on suggestions made by Monique in the document "Paper Structure building on the ideas we discussed earlier" in "Part 4. Further use of data", from March 2023). So the item group *Governance* is made up of 19 items. The item group *Outcomes* is made up of the three items in the `Outcomes (O)` category. Table \@ref(tab:itemstab) shows all items going into the IRM.

```{r itemstab}
irm_items %>% 
  select(item = dimension, tier1)%>% 
  mutate(
         itemtype = c(rep("Governance", 19), rep("Outcome",3))) %>% 
  kable(
    format = "html",
    caption = "Items in the IRM, their original Tier 1 DF category, and their item type, as defined in the model.") %>% 
  kable_styling()
```

The distribution of responses in these items is displayed in Table Table \@ref(tab:itemstabdist).

```{r itemstabdist}
irm_items_dist %>% 
  kable(
    format = "html",
    caption = "Items in the IRM and their distribution of responses. Keep in mind that binary items were re-coded as 1 (low value) and 3 (high value), such that they always lack responses on the middle option.") %>% 
  kable_styling()
```

The first model is formulated as follows (only in terms computer code for now):

```{r}
fit$formula
```

... with regularizing priors:

```{r}
prior_summary(fit)
```

The model summary is:

```{r}
summary(fit)
```

Convergence was satisfactory (judging from R hat values and trace plots), but might be improved further with other configurations of the MCMC sampler of stronger priors. Posterior predictive checks look fine. As expected, extreme response distributions (e.g., all 12 villages scoring the same on an item) are not reproduced by the estimated model with the same sharpness -- this is due to the regularization induced by the priors and is desirable in our spare data context. An example are the posterior predictive checks in Figure \@ref(fig:ppc).

```{r ppc, fig.align='center', fig.cap='Predicted vs observed counts of responses grouped by item (left) and village (right).', out.width="100%"}
include_graphics("outputs/ppcs.png")
```

Coming to the estimated model parameters, Figures \@ref(fig:res1) and \@ref(fig:res2) visualize posteriors for the decomposition into "item effects" and "village effects" by the item response model, as well as posteriors for the discrimination parameter of each item.

```{r res1, fig.align='center', fig.cap='Posteriors with median and quantile-based credible intervals for (a) village effects on rating, (b) item effects on rating and (c) item effects on (inverse) discrimination. All parameters are on the log-odds scale.', out.width="90%"}
include_graphics("outputs/param_posts.png")
```

The model broadly reproduces what we knew from initial data analyses. For example, Kimana village scores low on governance and outcomes (Figure \@ref(fig:res1)a), ratings on external sanctions are very low (Figure \@ref(fig:res1)b) and of low discriminatory value (Figure \@ref(fig:res1)c) - both is of course a consequenceof them being $1$ for all villages (see Table \@ref(tab:sds)).

```{r res2, fig.align='center', fig.cap='The posterior correlation of Governance and Outcomes based on the IRM.', out.width="100%"}
include_graphics("outputs/cor_itm.png")
```

Figure \@ref(fig:res2) constitutes a first major result from item response modeling. \@ref(fig:res2)a shows that we can be quite certain that there is positive correlation between Governance and Outcomes across villages. In Figure \@ref(fig:res2)b, 500 posterior draws for the latent Governance and Outcome variable are drawn for each village and plotted as scatter. The lines show 90\% contours of the bivariate posterior for each village. We can see that (i) there is a positive association between the two latent traits, but that (ii) there is also a lot of uncertainty about the villages' positions in the multivariate space, (iii) especially for the Outcomes dimension (ellipses are "higher" than they are "wide"), which makes sense, since it is only based on 3 items (compared to > 10 items for the Governance dimension).

Of course, since all variables included in this model are from reports by the villagers themselves, partially from focus groups, there could be other (e.g., cognitive, social) reasons for the positive association of *Governance* and *Outcomes* items.

Note that the related association presented in Figure \@ref(fig:scattergso) during initial data analysis was stronger and more certain, which is (among other things) because we threw away variability by only looking at averaged scores and assuming these as known rather than as following a distribution based on unknown parameters.

*Potential ways forward after this section:*

- Note this model does not condition on any potential confounders, but merely describes an association. One could think about deriving adjustment sets with a principled approach like DAGs.

- Next steps might include analyses on the role of disturbances in the effectiveness of governance (rainfall; ethnic and economic heterogeneity; population growth), staying within above modeling framework

- We will also involve GIS-based outcome measures in a Bayesian modeling approach, where those will be not modeled as item responses. (It could be interesting to also relate interview/focus group responses on livelihood change with GIS data on crop land change; and to relate responses specifically to invasive species control with bare ground change.)

- There is also the question "Was NTRI a Success?", which would involve looking at GIS-based outcome measures for other villages / areas that were not part of the program. The feasibility and validity needs to be checked.

## Model bare ground change

We have "objective" outcome data, namely the GIS-based bare ground estimates for the village areas. There is one estimate for the average bare ground cover in "pre" intervention period (2010-2015), and one for the intervention period (2016-2021). A *high* amount in bare ground can be viewed as *unfavorable* commons outcome, limiting the available collective rangeland. Increases (or decreases) in bare ground can therefore be an indicator for commons governance quality and ultimately the success of the NTRI's project which had the goal to maintain health and connectivity of rangelands. Therefore, we want to analyze its relation to the latent *Governance* scores, for which we obtained posterior distributions in the previous section. However, it is important that a physical indicator like the amount of bare ground is also influenced by other factors rather than governance alone and is therefore a less direct (though more objective) indicator of governance outcomes than the self-reports by villagers which were summarized in the *Outcomes* items (e.g., the compliance with rules they observe). 

The DAG in Figure \@ref(fig:dag) visualizes our working causal assumptions.

```{r dag, fig.align='center', fig.cap='DAG for the effect of governance on bare ground cover.', out.width="85%"}
dag <- dagify(barePost ~ barePre + rain + inv + cropPost + gov,
              cropPost ~ cropPre + gov + rain,
              inv      ~ gov,
              rain     ~ u1,
              gov      ~ u1,
              inv      ~ u2,
              gov      ~ u2,
              barePre  ~ u3 + cropPre,
              gov      ~ u3,
              cropPre  ~ u4,
              gov      ~ u4,
              exposure = "gov",
              outcome = "barePost",
              latent = c("u1","u2","u3","u4"),
              labels = c("inv" = "invasives",
                         "barePost"="barePost",
                         "cropPost"="cropPost",
                         "barePre"="barePre",
                         "cropPre"="cropPre",
                         "rain"="rainfallPost",
                         "gov"="governance",
                         "u1"="U1",
                         "u2"="U2",
                         "u3"="U3",
                         "u4"="U4")
)
ggdag(dag, layout = "circle",
      text = FALSE, use_labels = "label",
      edge_type = "diagonal") + 
  theme_dag_blank(14)
```

We are interested in the direct effect of the latent governance trait of a village (*governance*) on bare ground cover after the intervention (*barePost*), but, given this DAG, a marginal measure of association will not estimate such an effect. Instead, what we encode in and learn from this DAG can be summarized as follows:

- Besides the direct effect of *governance* on *barePost*, there are two indirect effects, one via *invasives* (since good commons governance would limit prevalence of invasives, and less invasives could mean more bare ground), and one via *cropPost* (since good commons governance would limit agricultural conversion, and less crop land could mean more bare ground; this is also a variable based on GIS measurements). Both of these indirect paths hence work **against** the direct governance effect for this GIS-based outcome measure and need to be blocked

- The crop land cover (*cropPost*) and bare ground cover (*barePost*) in the intervention period are both affected by the cover in in the previous period (*cropPre* and *barePre*), as, e.g., more bare ground pre intervention likely also means more bare ground after intervention

- Given the largely uncontrolled design of the project, we can expect a host of unmeasured confounding influences (*U1* to *U4*), between governance quality and factors relevant for land use measurements. Of course, the quality of governance cannot be assigned randomly. For example, data explorations as in Figure \@ref(fig:pca) already showed that rainfall in the post intervention period and DF variables on `Governance Systems (GS)`, `Interactions (I)` and `Actors (A)` are negatively associated, which is likely due to influences of unobserved third variables related to geography (such influences would fall into node *U1*). Another example would be factors captured by node *U3*, which are village-level characteristics that can affect both how much bare ground there is in general (i.e., at baseline) and how well resource governance functions. Again, one can think of geography and all that comes with it (landscape, biodiversity, local politics, ethnicity, ...). Such unobserved confounders should be controlled for by using the observed variables they run through on their path to the outcome.^[These considerations would provide also a plausible explanation why there was no clear association found between average `Governance Systems (GS)` score and bare ground change in Figure \@ref(fig:scattergis) in the absence of any adjustment (for example, The "outlying" observation for the village Matale A in Figure \@ref(fig:scattergis), which has a "too bad" bare ground outcome for its high average `Governance Systems (GS)` score is among the villages with least rainfall, see Figrue \@ref(fig:vilrain); after residualizing on rainfall, the negative association emerges and the degree to which Matale A is an outlier is reduced, as is seen in Figure \@ref(fig:scattergisres)).]

Using the DAG to arrive at an adjustment set, we get the following variables to adjust for:

- *rainfallPost*
- *invasives*
- *barePre*
- *cropPost*

Note that this is based on the assumption that there are no further variables which directly or indirectly influence both governance quality and bare ground cover in our setting and whose influence is not yet blocked by above adjustment set. This assumption is likely false and hence we should understand the result as a principled attempt to arrive at an approximate estimate in an imperfect study design rather than definitive truth. **In any case, it would be worthwhile to think of additional factors that we should try to adjust for.**

The Gaussian model we fit is formulated as follows (only in terms computer code for now):

```{r}
fit2$formula
```

... with priors:

```{r}
prior_summary(fit2)
```

Importantly, we supply the posterior distributions for the village-level governance latent traits that were estimated in the IRM above (i.e., the distributions visualized along the **horizontal axis** in Figure \@ref(fig:res2)b), rather than discarding the uncertainty about the governance quality in the villages, for example by only using the posterior mean as predictor. This is done via a measurement error model using `brms::me()`, assuming Normal distributions. Note that in our special case we know that the true village *Governance* scores (i.e., if there was no measurement error of the latent trait, measured via the item responses) have mean = 0 and SD = 1, which we supply to the model via constant priors. We know this since this were parametrizing assumptions that went into the initial IRM (see above).

The model summary is:

```{r}
summary(fit2)
```

We see a plausible estimate for average rainfall (standardized to have SD = 1) in the post period: The more rain, the less bare ground. For the latent governance quality parameter in the model, we get a largely negative posterior, indicating that better governance leads to less bare ground, though we cannot be fully certain about the direction of this effect (see Figure \@ref(fig:scattergis)).

```{r megov, fig.align='center', fig.cap='Posterior for the effect of Governance quality, as measured via the DF variables, on GIS-measured bare ground during and after the programme.', out.width="65%"}
include_graphics("outputs/megov.png")
```

We can summarize our two models' results from the previous and the current section:

<style>
div.blue {background-color:#e6f1fa; border-radius: 5px; padding: 15px;}
</style>
<div class = "blue">

- Using a multivariate item response model, we find that there is likely a positive association between commons governance processes as measured with Ostrom's Diagnostic Framework (DF) variables and commons outcomes as measured with DF variables among the NTRI villages.
- The village-level latent trait of commons governance quality estimated in the item response model is negatively associated with bare ground cover change in the years during and after the NTRI prject (2016-2021), indicating better rangeland conditions. This is conditional on rainfall and two other potentially positive effects of rangeland governance that could however have the potential to negatively affect bare ground cover: control of invasives (as measured in the respective DF variable `O2.3: Invasives`) and control of agricultural conversion (measured as crop land via GIS).

</div>

<br>


*Potential ways forward after this section:*

- Is the DAG plausible? Should some arrows be added or removed? Should any more variables from the many data sources absolutely be accounted for (e.g., the external threats data table, the household survey data, ...). To see which we data we have, remember you can look here: <https://stefgehrig.shinyapps.io/TNC-NTRI/>

- Next steps might include analyses on the role of disturbances in the effectiveness of governance (e.g., ethnic and economic heterogeneity or population growth) -- if possible in a valid way -- staying within above modeling framework

- There is also the question "Was NTRI a Success?", which would involve looking at GIS-based outcome measures for other villages / areas that were not part of the program. The feasibility and validity needs to be checked.

<!-- ### Moderation of the Governance - Outcome association -->

<!-- We can check whether the positive relationship between the village-level *Governance* and *Outcome* latent traits, as defined any analyzed in the previous section, is affected by different types of *Disturbances*. We use following measured variables to proxy for different types of disturbances on the village level: -->

<!-- - Rainfall (average of mean annual precipitation 2016-2020) -->
<!-- - Population growth (annual relative increase from 2012 to 2020) -->
<!-- - Economic heterogeneity (SD of first PCA for toilet, wall, floor, roof, pikipiki, phones) -->

<!-- We fit almost the same model as in the previous section once for each disturbance covariate, but allow that the covariate can shift the mean responses differently for both item types, i.e., we add an interaction term.  -->

<!-- ## Conclusions / Recommendations -->

<!-- <style> -->
<!-- div.blue {background-color:#e6f1fa; border-radius: 5px; padding: 15px;} -->
<!-- </style> -->
<!-- <div class = "blue"> -->

<!-- - This is my first conclusion -->
<!-- - This is my second conclusion -->

<!-- </div> -->

<!-- <!-- END OF --> 
<!-- <br> -->

<!-- *** -->
<!-- <!-- SECTION --> 

## References
